backend/requirements.txt
fastapi
uvicorn
gunicorn
pymongo
psycopg2-binary
